db 모델 정의
