+++ backend/models/__init__.py
@@ -1,10 +1,11 @@
 """ORM 모델 패키지."""
 
-from .entities import Membership, Organization, Workspace, WorkspaceType
+from .entities import Membership, Organization, Workspace, WorkspaceType, User
 
 __all__ = [
     "Membership",
     "Organization",
     "Workspace",
     "WorkspaceType",
+    "User",
 ]
